chore: change the request validation place
As I'm going to need request validation in other places, it doesn't make sense to create a function attached to every input struct and repeat the same code over and over again. I moved the validation to a separated package (utils), receiving a generic struct in the argument to reuse in other handlers. The handlers just need to use the validator tags and send it to the function.